+++ app/controllers/admin/suppliers_controller.rb
@@ -1,37 +1,33 @@
 class Admin::SuppliersController < AdminController
-  before_action :find_supplier, only: %i[show edit update show_tasks show_users]
-  before_action :load_frameworks, only: %i[show show_tasks show_users]
+  before_action :find_supplier, only: %i[show edit update show_tasks show_users show_frameworks]
+  before_action :load_frameworks, only: %i[show show_tasks]
 
   def index
     @suppliers = Supplier.search(params[:search]).page(params[:page])
   end
 
   def show_tasks
-    @tasks = @supplier.tasks.includes(:framework,
-                                      active_submission: :files).order(due_on: :desc).page(params[:task_page]).per(12)
+    @tasks = @supplier.tasks.includes(:framework, active_submission: :files)
+                      .order(due_on: :desc, id: :desc).page(params[:task_page]).per(12)
     @tasks = @tasks.where(framework_id: params[:framework_id]) if params[:framework_id]
-
-    respond_to do |format|
-      format.html
-      format.js
-    end
   end
 
   def show_users
     @users = @supplier.users.page(params[:user_page]).per(12)
 
-    filter_status params[:status] if params[:status]
+    filter_user_status params[:status] if params[:status]
+  end
 
-    respond_to do |format|
-      format.html
-      format.js
-    end
+  def show_frameworks
+    @agreements = @supplier.agreements.includes(:framework).page(params[:framework_page]).per(12)
+
+    filter_framework_status params[:status] if params[:status]
   end
 
   def show
     show_tasks
     show_users
-    @agreements = @supplier.agreements.includes(:framework).page(params[:framework_page]).per(12)
+    show_frameworks
 
     respond_to do |format|
       format.html
@@ -61,14 +57,21 @@ class Admin::SuppliersController < AdminController
   end
 
   def load_frameworks
-    @tasks = @supplier.tasks.includes(:framework, active_submission: :files)
+    @tasks = @supplier.tasks
     @frameworks = @tasks.collect(&:framework).uniq.sort_by(&:short_name)
   end
 
-  def filter_status(status_param)
+  def filter_user_status(status_param)
     return if status_param.size == 2
 
     @users = @users.active if status_param.include? 'active'
     @users = @users.inactive if status_param.include? 'inactive'
+  end
+
+  def filter_framework_status(status_param)
+    return if status_param.size == 2
+
+    @agreements = @agreements.active if status_param.include? 'active'
+    @agreements = @agreements.inactive if status_param.include? 'inactive'
   end
 end